+++ src/main/java/racing/ConsoleApplication.java
@@ -1,6 +1,6 @@
 package racing;
 
-import racing.controller.RacingController;
+import racing.controller.ConsoleRacingGameController;
 import racing.domain.CarFactory;
 import racing.domain.Cars;
 import racing.ui.input.InputView;
@@ -9,17 +9,12 @@ import racing.ui.output.OutputView;
 import java.util.Arrays;
 import java.util.List;
 
-/**
- * @author 베베
- * @version 1.0.0
- * @since by 베베 on 2023. 04. 14.
- */
-public class Application {
+public class ConsoleApplication {
 
     public static void main(String[] args) {
-        RacingController racingController = new RacingController(getCars());
+        ConsoleRacingGameController consoleRacingGameController = new ConsoleRacingGameController(getCars());
         int count = Integer.parseInt(InputView.inputCount());
-        racingController.start(count);
+        consoleRacingGameController.start(count);
     }
 
     private static Cars getCars() {
