+++ src/main/java/racing/NumberGenerator.java
@@ -1,11 +1,5 @@
 package racing;
 
-/**
- * @author 베베
- * @version 1.0.0
- * @since by 베베 on 2023. 04. 18.
- */
-
 @FunctionalInterface
 public interface NumberGenerator {
