+++ src/main/java/racing/dao/CarDao.java
@@ -1,14 +1,10 @@
 package racing.dao;
 
+import org.springframework.stereotype.Repository;
 import racing.controller.dto.request.CarRequest;
 
 import java.util.List;
 
-/**
- * @author 베베
- * @version 1.0.0
- * @since by 베베 on 2023. 04. 18.
- */
 public interface CarDao {
 
     void saveCar(CarRequest request);
