+++ src/main/java/racing/dao/GameDao.java
@@ -1,12 +1,7 @@
 package racing.dao;
 
-import racing.controller.dto.request.CarRequest;
+import org.springframework.stereotype.Repository;
 
-/**
- * @author 베베
- * @version 1.0.0
- * @since by 베베 on 2023. 04. 18.
- */
 public interface GameDao {
 
     Long saveGame(int count);
