+++ src/main/java/racing/dao/RacingCarDao.java
@@ -7,12 +7,6 @@ import racing.controller.dto.request.CarRequest;
 
 import java.util.List;
 
-/**
- * @author 베베
- * @version 1.0.0
- * @since by 베베 on 2023. 04. 18.
- */
-
 @Repository
 public class RacingCarDao implements CarDao {
